Fix storybook
by removing dependencies of `WalletContext` from `Header` and from
`,storybook/preview.tx`.

Same for `MidgardContext`, which did not causes any issues in
storybook, but we don't want make any server-side calls there.
