+++ .storybook/preview.tsx
@@ -7,9 +7,7 @@ import { withThemes } from '@react-theming/storybook-addon'
 import { BrowserRouter as Router, Route } from 'react-router-dom'
 
 import { ConnectionProvider } from '../src/contexts/ConnectionContext'
-import { WalletProvider } from '../src/contexts/WalletContext'
 import { ThemeProvider } from '../src/contexts/ThemeContext'
-import { MidgardProvider } from '../src/contexts/MidgardContext'
 import { I18nProvider } from '../src/contexts/I18nContext'
 import { AppWrapper } from '../src/App.style'
 import GlobalStyle from '../src/Global.style'
@@ -21,16 +19,12 @@ addDecorator(withKnobs)
 
 const providerFn = ({ theme, children }) => (
   <ConnectionProvider>
-    <MidgardProvider>
-      <WalletProvider>
-        <I18nProvider>
-          <ThemeProvider theme={theme}>
-            <GlobalStyle />
-            <AppWrapper>{children}</AppWrapper>
-          </ThemeProvider>
-        </I18nProvider>
-      </WalletProvider>
-    </MidgardProvider>
+    <I18nProvider>
+      <ThemeProvider theme={theme}>
+        <GlobalStyle />
+        <AppWrapper>{children}</AppWrapper>
+      </ThemeProvider>
+    </I18nProvider>
   </ConnectionProvider>
 )
