--- src/components/header/Header.stories.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-import { storiesOf } from '@storybook/react'
-
-import Header from './Header'
-
-storiesOf('Components/Header', module).add('default', () => {
-  return <Header />
-})
