+++ src/const.ts
@@ -1,14 +1,11 @@
-import * as path from 'path'
-
 import { getAssetFromString, assetToBase, assetAmount } from '@thorchain/asgardex-util'
-import { remote } from 'electron'
 
 import {
   PricePoolCurrencySymbols,
   PricePoolCurrencyWeights,
-  PricePoolAssets,
   PoolAsset,
-  PricePool
+  PricePool,
+  PricePoolAssets
 } from './views/pools/types'
 
 // Rune ticker as const - just because we use it almost everywhere ...
@@ -40,13 +37,5 @@ export const RUNE_PRICE_POOL: PricePool = {
   poolData: { assetBalance: ONE_ASSET_BASE_AMOUNT, runeBalance: ONE_ASSET_BASE_AMOUNT }
 }
 
-export const APP_NAME = remote?.app?.name ?? 'ASGARDEX'
-
 // Whitelist of pools for pricing things
 export const PRICE_POOLS_WHITELIST: PricePoolAssets = [PoolAsset.BTC, PoolAsset.ETH, PoolAsset.TUSDB]
-
-export const APP_DATA_DIR = path.join(
-  remote?.app?.getPath('appData') ?? './testdata', // we can't access remote.app in tests
-  APP_NAME
-)
-export const STORAGE_DIR = path.join(APP_DATA_DIR, 'storage')
